requirements.txt
asgiref==3.2.10
boto3==1.14.28
botocore==1.17.28
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.42.0
django-boto==0.3.12
django-countries==6.1.2
django-crispy-forms==1.9.2
django-jquery==3.1.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
stripe==2.49.0
urllib3==1.25.10
